build: add WDK10 MSBuild project for Win7 virtio-snd
- Add a WDK10 (WindowsKernelModeDriver10.0) WDM driver project for Win32/x64 Debug+Release\n- Force /W4 + warnings-as-errors and wire up include/\n- Set TargetName=virtiosnd so MSBuild outputs virtiosnd.sys for the existing INF\n- Document MSBuild/ci/build-drivers.ps1 as the supported build path; keep build.exe files as legacy\n- Adjust a few casts/DbgPrint arguments so the skeleton stays /W4-clean

+++ drivers/windows7/virtio-snd/src/driver.c
@@ -165,7 +165,7 @@ VirtIoSndAddDevice(PDRIVER_OBJECT DriverObject, PDEVICE_OBJECT PhysicalDeviceObj
         FALSE,
         &deviceObject);
     if (!NT_SUCCESS(status)) {
-        VIRTIOSND_TRACE_ERROR("IoCreateDevice failed: 0x%08X\n", status);
+        VIRTIOSND_TRACE_ERROR("IoCreateDevice failed: 0x%08X\n", (ULONG)status);
         return status;
     }
 
@@ -274,7 +274,7 @@ VirtIoSndDispatchPnp(PDEVICE_OBJECT DeviceObject, PIRP Irp)
 
         status = VirtIoSndForwardIrpSynchronously(dx->LowerDeviceObject, Irp);
         if (!NT_SUCCESS(status)) {
-            VIRTIOSND_TRACE_ERROR("Lower driver failed START_DEVICE: 0x%08X\n", status);
+            VIRTIOSND_TRACE_ERROR("Lower driver failed START_DEVICE: 0x%08X\n", (ULONG)status);
             IoReleaseRemoveLock(&dx->RemoveLock, Irp);
             return VirtIoSndCompleteIrp(Irp, status, 0);
         }
